services/notification/requirements.txt
grpcio>=1.60,<2.0
protobuf>=6.0,<7.0
pika>=1.3,<2.0
jinja2>=3.1,<4.0
azure-communication-email>=1.0,<2.0
redis>=5.0,<6.0
